ChabotAPI/requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
dialogflow==1.1.0
